static link

+++ ApplicationManager/Application.cpp
@@ -247,7 +247,7 @@ void Application::spawnProcess()
 	std::lock_guard<std::recursive_mutex> guard(m_mutex);
 	long gid, uid;
 	Utility::getUid(m_user, uid, gid);
-	int cmdLenth = m_commandLine.length() + ACE_Process_Options::DEFAULT_COMMAND_LINE_BUF_LEN;
+	size_t cmdLenth = m_commandLine.length() + ACE_Process_Options::DEFAULT_COMMAND_LINE_BUF_LEN;
 	int totalEnvSize = 0;
 	int totalEnvArgs = 0;
 	Utility::getEnvironmentSize(m_envMap, totalEnvSize, totalEnvArgs);
